backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic-settings>=2.6.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
dashscope>=1.20.0
python-multipart>=0.0.9
langchain-text-splitters>=0.3.0
langchain-qdrant>=0.1.0
qdrant-client>=1.7.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
greenlet>=3.0.0
redis>=5.0.0
